Update patch set 3
Patch Set 3: Code-Review+1

(6 comments)

Overall I like it. But I wonder if we can simplify it by not managing ansible at runtime. Anyway, mostly just questions. I think if we decide to go forward, more explanation of why we feel that we need to manage ansible should be added to the spec.

Patch-set: 3
Label: Code-Review=+1