Report retried builds via sql reporter.
Since we added those to the MQTT reporter, we should also store them in
the SQL database. They are stored in the zuul_build table and can be
identified via the new "final" column which is set to False for those
builds (and True for all others).

The final flag can also be used in the API to specifically filter for
those builds or remove them from the result. By default, no builds are
filtered out.

The buildset API includes these retried builds under a dedicated
'retry_builds' key to not mix them up with the final builds. Thus, the
JSON format is equally to the one the MQTT reporter uses.

For the migration of the SQL database, all existing builds will be set
to final.

We could also provide a filter mechanism via the web UI, but that should
be part of a separate change (e.g. add a checkbox next to the search
filter saying "Show retried builds").

Change-Id: I5960df8a997c7ab81a07b9bd8631c14dbe22b8ab