Report retried builds via sql reporter.
Since we added those to the MQTT reporter, we should also store them in
the SQL database. They are stored in the zuul_build table and can be
identified via the new "final" column which is set to False for those
builds (and True for all others).

To not spoil the build results page in zuul web with those RETRY
results, we filter them out by default (final=false). Although, if
interested one could retrieve them via the web API by setting the
"final" parameter to true. If no final parameter is provided in the web
api, all builds are returned.

For the migration of the SQL database, all existing builds will be set
to final.

As I know already some people who are interested in seeing those builds
also in the UI, we could add e.g. a checkbox next to the search filter
saying "Show retried builds".

Change-Id: I5960df8a997c7ab81a07b9bd8631c14dbe22b8ab