Update patch set 1
Patch Set 1:

If we don't want to switch this to a dictionary, we could update zuul_return to automatically extend the artifact list itself.  It would be the only place in Zuul that extends a list when performing data structure merging.  I think I'd prefer to keep the behavior consistent, but if we really want, we *can* make zuul_return special.  This part of it doesn't have to deal with general data -- it's very specific input to Zuul.

Patch-set: 1
Label: Verified=0