[WIP] - URLTrigger based on the timer trigger
Idea: Better to implement just a *require* driver that could
be associated with the timer trigger.

The change propose an extension of the timer driver
that take an url in addition to the time trigger value.

The ETag header of the url is checked to a previous
value (from a memory cache) at each time triggering
and if the ETag has changed the event is sent to the
scheduler.

This could be a solution to implement an URL trigger
like the Jenkins one (URLTrigger Plugin).

Change-Id: I81b50a8c1b72cddca78b75c2203fa7a3b4e55bab