Propose the URL driver to implement an URL change Require Filter
This change propose a new 'url' driver. The driver is minimal and
only implements a require Filter. It aims to be used with the timer
trigger as a require filter to check if an url header has changed
since the last check (that is the previous timer event).

This could be improved to check more than an header change.

Example of pipeline:

- pipeline:
    name: artifact-checker
    manager: independent
    trigger:
      timer:
        - time: '0/5 * * * *'
    require:
      url:
        url: https://domain.com/resources
        attr: 'Etag'
        connection: gerrit
    success:
      sqlreporter:
    failure:
      sqlreporter:

Change-Id: Iccba6741523845a3f72dcbb4c8e2cd4050572509