URLTrigger driver time based - artifact change jobs triggering driver
The aims of the driver is to provide a solution for triggering
jobs based on an artifact state change.

Given an URL and an attribute (an HTTP header field name) the
URLTrigger is able to discover that the external resource has
evolved since the last check. The delay attribute is used
to set the check state rate.

Response headers such as 'Last-Modified' or 'Etag' could be used
with that driver to trigger jobs based on the URL content change.

This driver is a proposal to bring a solution similar to
the URLTrigger plugin of Jenkins.

- pipeline:
    name: foo-artifact-changed
    post-review: true
    description: Jobs in this queue are triggered when the foo artifact has changed.
    manager: independent
    precedence: low
    trigger:
      url:
        - delay: '*/10 * * * *'
          url: https://domain.com/artifacts/foo/2.x/foo.tgz
          attribute: ETag
    success:
      sqlreporter:
    failure:
      sqlreporter:

- project:
    name: myproject
    foo-artifact-changed:
      jobs:
        - functional-test-1
        - functional-test-2

Change-Id: I58ed1fb4bc42575400110399dadb26e797a7f4d8