Update patch set 2
Patch Set 2:

Thanks for the feedback. There are currently three implementation of that (this one is the first). I'm trying to figure out the best way to integrate that feature in the Zuul workflow. I agree that Etag could be the default and the field to check should be configurable. See other implementation:

- https://review.openstack.org/#/c/635554/
- https://review.openstack.org/#/c/635567/ (this one sounds better to me)

Patch-set: 2