Update patch set 5
Patch Set 5:

We can have a web trigger, but we can't have a job filter.  That undermines the pipeline model in Zuul, where pipelines implement workflow processes.

To implement your example of a deployment job, you could create a deployment pipeline, add a project to it with a single deployment job, and add the web trigger to it.  You could also add a post-merge trigger to it as well, so then deployment happens every time a change lands, or also upon request via the web trigger.

Patch-set: 5