WIP: Symlink ansible plugins
TODO: This change needs to be squashed into its parent but is kept
separately for easier review at the moment.

For better maintainability the ansible plugins live in
zuul/ansible/base where plugins can be kept in different versions if
necessary. For each supported ansible version there is a specific
folder that symlinks the according plugins.

Change-Id: I411b52f24ea9c56bcd46546995e9e004cab33db7