Fix excluded config if excluded in earlier tenant
Consider the following scenario where tenant-one includes org/project1
without any config. Zuul currently doesn't load the config of
org/project1 in tenant-two as well.

- tenant:
    name: tenant-one
    source:
      gerrit:
        untrusted-projects:
          - org/project1:
              include: []

- tenant:
    name: tenant-two
    source:
      gerrit:
        untrusted-projects:
          - org/project1

This is caused because we cache an empty config if we exclude all
config from the repo the same way as if the repo wouldn't have any
config. This causes the cat job to be skipped when loading
tenant-two. This can be solved by marking the empty config that will
be cached as 'this has no config because it has been skipped'. Then we
can detect that case and still trigger a cat job if we need the config
in a later tenant.

Change-Id: I1e7e68f11ca4fb7d1e52eee6d4c9a9d8331b012c