Add new merger job to get role definitions from a repository
Currently, Zuul does not provide any information about the Ansible roles
used within the system.
If we want to retrieve those information in Zubbi directly from Zuul, we
could therefore use a merger job that extracts the necessary data from
the repositories itself.

This merger job could also be used from Zuul web to list the roles or
provide an API similar to the jobs endpoint.

Change-Id: I1af0464ef3638f12046a8c02a8303c147d31ee19