Update patch set 1
Patch Set 1:

Regarding the python runtime container idea, I like it but it's less trivial as on the first glance because we also need to install bindep requirements and run it privileged because otherwise the tests won't work because of bubblewrap. Because of this it might make sense to do this in a separate change and maybe even in a separate docker(-compose) file.

Patch-set: 1
Label: Verified=0