Update patch set 11
Patch Set 11:

The change looks fine, but I still have concerns about the approach. I'm not sure if it it is a good user experience to have the inventory changed by the job after it has been constructed by nodepool and zuul. It feels too much like a side-affect and would require the user to be paying special attention to any jobs that may be modifying the inventory (even if only to add new kubectl hosts).

I don't feel strongly enough either way to leave a vote.

Is there a better way for users to interact with the kubectl host without adding it to zuul's inventory. Maybe running ansible from within ansible as a job with its own inventory for example?

Patch-set: 11