Update patch set 11
Patch Set 11:

The main use-case is for Zuul deployement with only kubernetes. In that situation, trusted phase could take care of creating the pods using executor's localhost kubectl command, so that untrusted job could run without privileged in kubernetes.

Another solution would be to dump the pod name in the job's log so that untrusted job could lookup the information and use add_host, but that makes the user's test playbook more complex.

Patch-set: 11
Label: Verified=0