Switch to LRU + TTL based sha to PR cache
This updates the github driver to cache PRs by sha using cachetool's
TTLCache. This is an LRU cache with time based expiration of objects as
well.

We make this change because we need to cache closed PRs so can't rely on
the action of closing a PR to remove the PR from the cache. Since we
don't have a good method of evicting entries we fall back to LRU and
time based expiry.

Change-Id: I5fb6c8b33f9eed221a8b84e537f02e7dccf2d2df