Update patch set 1
Patch Set 1:

What do you think about adding a timer that tracks the time from event received until the driver hands it over to the scheduler? That way we can track the delay introduced by the event preprocessing in the github driver. This might also help in the future to judge on future improvements such as using graphql or similar.

Patch-set: 1