Update patch set 4
Patch Set 4:

I think that basing this on a number of available CPUs is not a best option. Cloning git repositories appears to be IO-bound in my (limited) testing.

Here's a bunch of numbers that I'm getting on my installation. All of Zuul is on one VM with latest centos 7, two virtual CPU cores, 4GB RAM, hosted on a company-internal vmware cluster which is supposed to be backed by an SSD disk array, but definitely not NVMe- or PCIe-class flash; not sure if it's SAS or SATA and no idea about the actual storage paths. Python 3.6 from SCL.

I'm cloning 150+ repositories, and the times are between 'Beginning job' and 'Git updates complete' in my executor-debug.log:

- zuul release 3.5.0: 85s (run #1), 85s (run #2), 87s (run #3)
- this change on top of current master, i.e., two threads: 53s, 42s, 57s
- this change modified to use just one thread: 109s, 85s, 85s
- this change, but using 8*cpu_count (=16) threads: 19s, 18s, 17s
- 32*cpu_count (=64) threads: 21s, 21s, 19s

During all this testing, my Gerrit settings were these, meaning that Gerrit was actively throttling the number of connections to be used for fetch:

 [sshd]
   threads = 8
   batchThreads = 2
   commandStartThreads = 2

So this is a step in right direction, but I would recommend increasing the thread count even more. If that is not possible for some reason, please consider adding a config option to set the thread count manually.

There's one more step which takes about 70s for me: the subsequent clone from the just-updated repositories to a temporary location for one job run. This looks like something which is triggered from AnsibleJob._execute() which calls merger.getRepo(). I will see if I can make that run in parallel. In the meanwhile, thanks for this nice improvement.

Patch-set: 4
Reviewer: Gerrit User 13413 <13413@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0