Update patch set 7
Patch Set 7: Code-Review-2

The quick-start test is for testing that the quick-start works.  The only actions it performs are actions that we tell people to do in the documentation.  I don't think zuul-runner is something that we would want to add to quick-start.  It should use the existing unit-test framework.

Patch-set: 7
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2