Update patch set 6
Patch Set 6:

The reason I am resistant to a job filter, and having triggers alter
variables, is that Zuul is designed as a gating and automation system.
It is designed to facilitate workflows and automation that are
git-driven.  We intentionally have not added features that allow users
to run individual jobs (with or without parameters), because it leads
to processes which are not automatable.  Taking your previous example
of a deployment -- it's perfectly fine for a deployment to be triggered
by a human, a URL changing, or a git commit merging.  Those different
triggers can all be attached to one or more pipelines to run a single
deployment job.  However, if one of the triggers allows a user to select
which job to run, then we've just created a situation where it may only
be possible to deploy the system by asking a user to run a job.

I will note that even the emergency admin tool which we use when a job
fails, "zuul enqueue-ref", will still run all of the jobs in a
pipeline.

Patch-set: 6