Request child nodes of paused jobs at higher priority
When requesting nodes for jobs which depend on paused parent jobs,
do so at a priority slightly higher than the pipeline priority
for that job.  E.g., if a gate pipeline priority is 200, then the
child job nodes will be requested at 199.

This is based on the theory that by the time the parent job has
run and paused, the queue item has already spent its time in line
waiting for nodes.  It's fair for it to get the additional
required nodes ASAP, and doing so means that the paused job may
finish quicker, reducing wasted node-time.

Change-Id: I5a210d37ef223e93361004a26ef327ffeb530a3c