Update patch set 1
Patch Set 1: Code-Review+1

This makes sense to me.

To be honest, as someone who is still learning Ansible, I had no clue that it is happy to interpret the template language even in content that's coming from external sources. In light of that, a base64 form will make it obvious to me that that particular variable "has to be sanitized" in order to pass through the YAML machinery. That way, it is more intuitive to me than a random YAML tag (and I had no clue that there is something like an "unsafe" tag, anyway).

I've only consumed stuff from zuul.* variables via ansible's to_json filter, so I am probably on the safe path already.

Patch-set: 1
Label: Code-Review=+1