Update patch set 11
Patch Set 11:

Since this change (and children) is a substantial change to the executor and user-visible interaction with the system, I'd like to make sure that we get reviews from at least the following folks:

* Clint
* Tobias
* Monty
* Myself

as well as Joshua (who is the original author of some of these changes, but I'd like to make sure he has a chance to review any updates).

That's more than we usually require for changes, but I think it's both warranted and beneficial for these changes.

Patch-set: 11
Label: Verified=0