Update patch set 1
Patch Set 1:

And what about post-run playbooks that run after this?  For instance, we recommend that folks run 'remove-build-ssh-key' in a base post playbook, but that won't work with this.  In practice, I think the only time it would be safe to use this is in a base-job post-playbook, when you know for certain that all on-host operations have completed.  However, consider that combined with my previous concern -- that means that all jobs would run this, and therefore no jobs would ever be subject to an autohold.

Patch-set: 1
Label: Verified=0