Update patch set 13
Patch Set 13: Code-Review-1

Wait, I'm confused.  This change creates freeze-jobs which produces an enormous amount of information about what jobs will be run, including playbooks, roles, secrets, etc.  It uses the model toJson methods to do this.

The next change in the series creates freeze-job which produces nearly the same information for an individual job, but instead of using the model toJson, it uses the executor serializer.

It seems to me that these two endpoints should either be producing the same information (ie, serialized job descriptions in the same format).

How would you explain the purpose and use of the two different endpoints?

Patch-set: 13
Label: Code-Review=-1