Implied branch matching must use fullmatch
The when not expressing branches each job automatically matches the
branch where it is defined. Currently that match is not a fullmatch.
This leads to the situation that when having a job on branch stable and
on stable-foo a change to stable-foo matches both because stable is a
submatch of stable-foo. Using full regex matching on implied branches
fixes this.

Change-Id: I37e2a5dcf5df5c85812327a09d2678c0544bd366