Install ansible during executor startup if needed
In order to make running zuul easier we want the possiblility that the
executor installs the supported ansible versions during startup. This
adds this functionality as well as a config switch to disable it and a
config option to optionally specify the install location. The default
location is <state_dir>/ansible-bin.

Change-Id: I1858e4fb40190626d001e20b48cf7e69ad35d634