Update patch set 4
Patch Set 4:

Thank you for the feedback Waldemar. The issue was that phase2 and phase3 job had a branch matcher and the implementation in PS3 incorrectly look for hierarchy in that case. In PS4 I removed branch filter in this page, as it is already the case. I think a follow-up should add tabs at the top of the list to show the job's hierarchy per branch.

Patch-set: 4
Label: Verified=0