Fix duplicate and reversed artifacts
The "current" provides/requires test (which is supposed to test
live changes) was inadvertently following the SQL code path in
some cases.  The test has been updated to include a third job
which continues running until the end of the test to ensure that
the changes remain in the pipeline the entire time.

It, and the SQL version, have been extended with one more change
to illustrate a bug where artifact data were being included twice.
This was because the data collection method has two recursion points:
a) if it is called on a non-live change and it finds a live version
   of the same change, it recursively calls itself on the non-live
   version of the same change.  Essentially, it "switches" to that
   version of the change and continues.
b) at the end of normal processing of a change, it recurses "up"
   the queue to the next change ahead.

The (a) case did not return in some cases, so in those cases, it
would "switch" to a change, add its data, recurse up the queue,
then, after unrolling back to the point of the switch, would then
recurse up the queue again in case (b).  The solution is simply
to return in all recursion cases, so only (a) or (b) happens for
any single change.

Finally, the artifacts were being returned in reverse order -- it
is more useful for them to be returned such that the latest version
of a given artifact is returned last so that competing versions
are overwritten in the correct order.

Change-Id: I41ac336c1bc776609645e3662003f2df076dd1d5