Update patch set 6
Patch Set 6: Code-Review+2

I love it.

Mentioned in IRC - but I think it might be a nice follow up to add a toggle to let people see a flat list as well. Go to the jobs list, search for "openstacksdk" - and while the results are super informative, if you're looking for a specific openstacksdk job that you can't remember the exact job name, it's harder to scan the list of them in tree form. (openstacksdk seems to be a good complex case) But if we did that followup, I think defaulting to tree is nice, so maybe the toggle could be "Flatten"?

Patch-set: 6
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2