authentication config: add optional token_expiry
The Zuul admin can configure authenticators with an optional
"token_expiry" field, which is the maximum age in seconds
for a valid authentication token. By default there is no
maximum age set for tokens, except the one deduced from
the token's "exp" claim.
If "token_expiry" is set, tokens without an "iat" claim will
be rejected.

This is meant as an extra security to avoid accidentally issueing
very long lived tokens through the CLI.

Change-Id: I9351ca016b60050b5f3b3950b840d5f719e919ce