Ensure correct lexical sorting of node requests
The recently introduced change to request child nodes of paused jobs
with a higher priority causes the priority to be lower in case the
pipeline precedence is set to 'high'.

Due to the lexical sorting of node requests in Nodepool, 99 will be
treated as the highest prio in this case.

This is especially apparent when Nodepool is up against quota.

Change-Id: I094dee4f357c9974b6d9e95fcd70b02115d9de93