Update patch set 2
Patch Set 2:

Using getSafeAttributes() here would require extending the Attributes class to not throw an exception in case a non-existent attribute is used in a format string. This is currently made sure in the reporter itself (getattr()).

What I don't quite like about using getSafeAttributes() is that it add just another layer of indirection for e.g. the tenant and pipeline since they only have a name attribute. Granted, there might be more attributes in the future.

I was also trying to align the available format fields between the MQTT and SMTP reporter. Since we need to maintain backward compatibility, we can't change this in the MQTT reporter. So the two reporters will diverge further.

Patch-set: 2
Label: Verified=0