Skip ansible validation for non-ansible tests
The non-ansible test cases don't need to validate if ansible is
installed correctly so skip that in this case. Also copying the
ansible module overrides can be skipped in this case.

This shaves off two seconds for each non-ansible test case and reduced
the runtime of the full test suite locally from 420s to 320s.

Change-Id: I9af3c932589ad10804ae0e8a9f2551e4fe35990e