A reporter for Elasticsearch
It has the capability to index build and buildset results in
an index.

With the help of tools like Kibana, advanced analytics dashboard
could be built on top of the zuul Elasticsearch index.

Simply define the Elasticsearch uri and index name, then add the
elasticsearch reporter to a pipeline.

Job's variables and zuul_return data are exported along with build
results under the job_vars and job_returned_vars fields.

Change-Id: I5315483c55c10de63a3cd995ef681d0b64b98513