A Zuul reporter for Elasticsearch
It has the capability to index build and buildset results.

With the help of tools like Kibana, advanced analytics
dashboard could be built on top of the Zuul Elasticsearch
index.

Optionally job's variables and zuul_return data can be
exported along with build results under the job_vars and
job_returned_vars fields.

Change-Id: I5315483c55c10de63a3cd995ef681d0b64b98513