Elasticsearch Zuul reporter
A reporter for Elasticsearch with the capability to index build and
buildset results in an index.

With the help of tools like Kibana, advanced analytics dashboard
could be built on top of the zuul Elasticsearch index.

Simply define the Elasticsearch uri and index name, then
add the elasticsearch reporter to pipeline. Optionnaly
job's variables can be exported along with build results.

Change-Id: I5315483c55c10de63a3cd995ef681d0b64b98513