Increase zk session timeout during tests
When running the full test suite the test cases are running in a
highly contended and variable environment. We see that the default
session timeout of 10 seconds during the tests is too low in some
cases and cause test failures. Increase that timeout to give zookeeper
more time to respond.

Change-Id: Icb2832008147ab344c60094af574db5c02218cf1