Fix test race in test_periodic_override
The test case test_periodic_override checks if project overrice in
timer triggered pipelines works. For this it enables the timer trigger
and then creates the branch stable/havana. After some time it checks
all in flight builds that they checked out the correct
revision. However the branch creation itself first creates a branch
and then a commit. If the timer kicks in between those two steps zuul
will use the same commit for both branches and the test fails.

While it's easy to fix that in the test itself by first creating the
branch and then enabling the timer we don't know if there are other
test cases that might hit this race. Therefore we should eliminate
this race while creating the branch. We can easily do this by first
creating the commit on a detached head and after this create the new
branch head directly on the correct revision.

Change-Id: I21866978f4ffaaad4484b2dc93c6a6afa0745f9d