Update patch set 22
Patch Set 22: Code-Review-1

Starting at about this change, I'm having trouble following the series.  I don't know what the final product is going to look like, so it's hard fo me to evaluate the changes.

I'm going to make some suggestions for documentation on this change and some others.  The documentation will help developers in the future and it will also help us review this now (nowadays, we would create a spec before starting work that is this complicated, but we don't have that).

You can either update this change to include the documentation, or you can make it as a followup change, but if you do that it would be nice to insert it into the stack before or after this one.

For this change: can you add inline documentation to the AnsibleJob and AnsibleJobBase classes and include those in a new page in the developers guide at https://zuul-ci.org/docs/zuul/developer/index.html ?  I'd like to understand what should go in AnsibleJobBase, what should go in AnsibleJob, and why.  I understand that this change moves some things into AnsibleJobBase, and later changes move more.  That's great -- this change should just be a move so that later substantial changes are easier to follow.  So the documentation can be a little bit ahead of the code if you want.

Patch-set: 22
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1