Update patch set 4
Patch Set 4: Code-Review-1

It looks like we're talking about a configuration file for the runner.  That surprises me.  I think that means we need documentation accompanying this change which describes how we expect users to run the program.  It sounds very inconvenient to require a config file.  Can we support command line options too?  Of course, if there are a lot, that's also inconvenient.  Therefore, can we minimize the options?  A document which explains all the options would make it easier to see what we're aiming for and what could be combined or removed.

Patch-set: 4
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1