Update cached repo during job startup only if needed
Currently we update each cached repo that is involved in a job during
the start phase of the job. After updating we clone each of these
repos into the work dir.  However when running a job we already got a
repo state containing all references we need. So we can check if the
cached repo already has everything we need. If the repo state doesn't
contain revisions we don't have yet we can skip updating the repo and
thus safe a fetch.

This can improve job startup times significally especially for
expensive jobs with many required projects.

Change-Id: I9364e438d581b068fa19c9dfc24adab60479c385