runner: implement prep-workspace
This provides the beginning for a runner which will be able to execute
AnsibleJob's in a context outside of the executor server.

It begins by providing functionality to prepare the repositories. This
will be exposed as its own command, so that a developer can run the
steps individually (and interact where they need).

Eventually the prepared repositories will be able to be ran with Ansible
via the zuul-runner.

This change previously held refactoring now in its own commit
(I15a92d3944eab1812b34e9ea1a2a6be9d65a4244).

Change-Id: I93fb1c0a50aa3871fa38470d08601d9bee372f2e
Co-Authored-By: Tristan Cacqueray <tdecacqu@redhat.com>