Update patch set 14
Patch Set 14:

(1 comment)

> lgtm, but obviously the openapi.yaml isn't complete. I suspect we
 > want that to be complete before publishing this as it will be
 > publicly viewable (unless we want to only publish a subset of API
 > endpoints to begin with).

I imagine we could either (a) highlight the fact that the spec in incomplete wherever we're publishing it or (b) split out the spec and documentation aspects of this, to allow us to iterate on the spec before we publish it.

It might also be worth looking at whether one could verify the API against the spec. We do that in Patchwork [1] and it helped find a whole lot of bugs with both the spec and the API. I could help out here...once I figure out where the API actually lives :) (relative URLs)

[1] https://github.com/getpatchwork/patchwork/tree/a4698b2ed/patchwork/tests/api

Patch-set: 14
Label: Verified=0