Move repository preparation into common class
This change moves code that is not unique to the executor server into
the common library for reuse elsewhere.

This code was previously a part of
I93fb1c0a50aa3871fa38470d08601d9bee372f2e but split out to help separate
out the refactoring from the implementation of a runner.

Co-Authored-By: Tristan Cacqueray <tdecacqu@redhat.com>
Change-Id: I15a92d3944eab1812b34e9ea1a2a6be9d65a4244