Update patch set 15
Patch Set 15: Code-Review-1

(1 comment)

-1 for the inline comment about override.

I still would like to explore more the idea of flattening the variables.  That would also help with the inline comment.  If we could store "job_vars" as one indexed field, and "job_result_vars" (or similar) as a second indexed field, that would make things much simpler and more future-proof.  I'm sure that would have an impact on some searches, but I don't know what.  Do fields need to be indexed in order to appear in searches?  Would the searches be sufficiently constrained by other indexed fields such that a filter on an unindexed field would be able to return the desired results in a reasonable time?

I think it would be good to expand a little bit on use cases so we know how to approach that decision.  Then it might become clear whether we should specify variables to add as indexed fields, or flat (json-encoded) fields, or if we should do a combination of both.

Patch-set: 15
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1