Add link to builds in job page
Since the introduction of the tree view of the jobs there is no
workflow anymore to directly get to the build results of this
job. Especially when working with periodic jobs it's now more
cumbersome to get to these result. Thus add a link to the build
results on the job page.

Change-Id: Ia5c9dcbf91bb606ce0bab0cfaea1f406e95c8630