Update patch set 1
Patch Set 1:

I think we should perhaps mention and evaluate a few possible alternatives. The one that I would like to give thought to is making artifacts more of a first class citizen in the zuul-gearman protocol.

Rather than relying on playbooks to create and upload a manifest, I would prefer to see jobs return a manifest to zuul. Zuul would then need to store this manifest itself, likely in the database. Some of the reasons would be:
 - it avoids having to worry about CORS,
 - the manifest could link to arbitrary url's, allowing artifacts on different endpoints/backends (granted, this could happen anyway, but it shouldn't be expected that all the artifacts are stored next to the manifest)
 - zuul could be responsible for high-level execution logs that it might want to provide/add into the manifest (for example, if a job doesn't have a playbook to upload anything, at the very least we could log along side the result what was actually called/processed by the executor (or maybe even the console-logs etc)).

Patch-set: 1