Update patch set 1
Patch Set 1:

We do already support returning artifacts as first class objects -- we just consider artifacts to be things like "tarballs", "images", "documentation builds", etc.  Not individual log files.  While I would love to store the log manifest in the database, I don't think we should.  The manifest may (and in OpenStack's case, will) contain thousands (perhaps tens of thousands) of files for every build.  Storing that kind of data in a database is of course regularly accomplished, but it may entail a level of management that I don't think we want to impose on Zuul admins.

Even if we did, it would not address your first reason: CORS.  That still needs to be considered when fetching the actual logs themselves for display in the web interface.  Even the current build page fetches the job-output.json file for display.  The proposed change extends that to other files as well.  In other words, the files subject to CORS would be:

* manifest.json
* job-output.json
* individual log files

The last two are not candidates for being in the database, so the problem stands regardless.

Regarding your next point, it is already assumed that artifacts may not be stored next to the manifest.  The docker image jobs have artifact links to a docker image registry, for example.
However, it is reasonable to assume that the *log file* manifest is stored with the logs.

And on the third point, I would like to explore storing some more data in the build database about catastrophic post-run failures, so that if a log upload fails, there is some recourse.  But I think we should plan on minimizing the use of that so that it doesn't drive scalabality issues with the database.  In other words, I think that should be more of a failsafe than something we count on for every build.

Patch-set: 1