Use apt mirror infrastructure during zuul-quick-start
The zuul-build-image job is potentially unstable because it doesn't
use the local mirror infrastructure. This introduces the build arg
PACKAGE_MIRROR that instructs the image build process to use the local
package mirror for apt. Note that mirror config for pip still needs to
be implemented in a later change.

Change-Id: Ib0db7b608e13cd533e151e71e4668c83df461182