Provide roles per tenant via Zuul API
This utilizes the merger job to retrieve the role information for all
projects in the tenant. This way we could use the Zuul API as common API
between Zuul and Zubbi.

NOTE: Using this API endpoint may take some time for larger
projects/tenants as the role information is directly read from the
repo's files via the merger job. We should think about a way to
cache/store those information in Zuul (or maybe there is already a
way?).

Change-Id: Iaeb98577555bc3870ac5a751cea2339611fddeda