gerrit: Add support for 'oldValue' comment-added field
Since 71b58990e927 "Append approval info to every comment-added
stream event and hook"[1], merged in 2.13[2] every comment-added
event contains the approval values of the change regardless of
whether the comment in question actually caused an approval
transition. This breaks the Gerrit approval trigger filter,
effectively causing every new comment on a change that has a given
approval to trigger as if the comment added that approval.

The Gerrit change added an 'oldValue' field to the event, only when
ther is a transition, so this can be used to go back to the old
behaviour - effectively we ignore any approval entry that doesn't
have this field.

Unfortunately this Gerrit API change is not backwards compatible (an
approval entry without an 'oldValue' field means a totally different
thing on new Gerrits than old ones) so we need to explicitly check
the Gerrit version to interpret it. This is done via an SSH
command. There is also an HTTP API endpoint for this but the SSH one
is better documented (and we don't have a HTTP GET function in
GerritConnection AFAICS). If we cannot parse the Gerrit version we
assume the newer behaviour; the rationale being that our failure to
parse is probably because the format has changed (it is the future)
or Gerrit was built from a custom SHA (the user is probably on the
bleeding edge).

[1]https://gerrit.googlesource.com/gerrit/+/71b58990e9278e0d1128c396dd382a9a7cd5c6c9
[2]https://www.gerritcodereview.com/2.13.html#misc

Change-Id: I7c5381d83860a7b5766558fce2907a663369e4f1