Update patch set 9
Patch Set 9:

Just read Jim's comment about "Keeping this on the long term" in the tenant/config change. I think, for the Zubbi API we should use something we could also keep for longer. So, if neither of the changes is appropriate because of the output size, we could also think about another way to provide the data.

Patch-set: 9