Update patch set 1
Patch Set 1: Code-Review-1

Not opposed to doing this (although I still think the uid thing is absurd security theatre that provides no value only pain. "Make sure files that need to be written to are group-owned and group-writable by root" ... TOTAL GARBAGE) But the fact that it's an absurdity doesn't mean anyone is going to change it, since some clearly clueless product managers have convinced themselves it's "security".

HOWEVER - I'm -1 because there is no license or copyright info on the file, and because looking at the link from where it came from there is no license information in that repo. Could we please get license information and add both a license and copyright line?

Patch-set: 1
Label: Code-Review=-1