Fix occasional gate resets with github
We faced some occasional gate resets without failed jobs. We tracked
this down to the following scenario. The first item in the gate is
successful and zuul reports and merges the pull request. It first
comments that the gate jobs succeeded and then merges the pull
request. If this is successful the github driver marks the item as
merged in zuul's internal data structure. After reporting the pipeline
manager checks if the merge was successful by checking this flag and
usually everything is ok. But sometimes an earlier event of that pull
request sneaks in and is processed just between the merge call and the
pipeline manager check. If that happens the pull request data is
re-fetched from Github which might return slightly outdated data and
resets the is_merged flag back to false. If that happens the pipeline
manager thinks something went wrong with merging (however the merge
was successful), resets the gate. Especially if there are several
items with many long running jobs behind this gate reset can be very
expensive and waste many resources.

Change-Id: I839f307cf9f55e7ee78de8c107770cda83d390b7